+++ lib/ResourceSet.h
@@ -26,7 +26,7 @@ namespace Res
 		DLL_LINKAGE ResourceSet operator OPSIGN(const TResource &rhs) const	\
 		{														\
 			ResourceSet ret = *this;							\
-			for(int i = 0; i < size(); i++)						\
+			for(size_t i = 0; i < size(); i++)						\
 				ret[i] = at(i) OPSIGN rhs;						\
 																\
 			return ret;											\
@@ -38,7 +38,7 @@ namespace Res
 		DLL_LINKAGE ResourceSet operator OPSIGN(const ResourceSet &rhs) const	\
 		{															\
 			ResourceSet ret = *this;								\
-			for(int i = 0; i < size(); i++)							\
+			for(size_t i = 0; i < size(); i++)							\
 				ret[i] = at(i) OPSIGN rhs[i];						\
 																	\
 			return ret;												\
@@ -71,7 +71,7 @@ namespace Res
 		DLL_LINKAGE int operator/(const ResourceSet &rhs)
 		{
 			int ret = INT_MAX;
-			for(int i = 0; i < size(); i++)
+			for(size_t i = 0; i < size(); i++)
 				if(rhs[i])
 					vstd::amin(ret, at(i) / rhs[i]);
 
@@ -80,7 +80,7 @@ namespace Res
 
 		DLL_LINKAGE ResourceSet & operator=(const TResource &rhs)
 		{
-			for(int i = 0; i < size(); i++)
+			for(size_t i = 0; i < size(); i++)
 				at(i) = rhs;
 
 			return *this;
